2.0.0-SNAPSHOT
-updated folder names
-updated package names
-updated pom.xml
