+++ pom.xml
@@ -12,7 +12,7 @@
   </parent>
   <groupId>com.github.jenkinsci</groupId>
   <artifactId>zaproxy</artifactId>
-  <version>2.0.0</version>
+  <version>2.0.0-SNAPSHOT</version>
   <packaging>hpi</packaging>
 
   <name>Official OWASP Zed Attack Proxy Jenkins Plugin</name>
