+++ src/main/java/com/github/jenkinsci/zaproxy/report/ZAPReportCollection.java
@@ -22,17 +22,15 @@
  * SOFTWARE.
  */
 
-package com.github.jenkinsci.zaproxyplugin;
+package com.github.jenkinsci.zaproxy.report;
 
 import java.io.Serializable;
+import java.util.HashMap;
+import java.util.Map;
 
-import org.kohsuke.stapler.DataBoundConstructor;
-
-import hudson.Extension;
-import hudson.model.AbstractDescribableImpl;
-import hudson.model.Descriptor;
-
-/*
+/**
+ * This class contains all ZAPreport instance of the application. It's a singleton class so the application contains only one instance of the class.
+ *
  * @author Goran Sarenkapa
  * @author Mostafa AbdelMoez
  * @author Tanguy de Lignières
@@ -40,48 +38,34 @@ import hudson.model.Descriptor;
  * @author Thilina Madhusanka
  * @author Johann Ollivier-Lapeyre
  * @author Ludovic Roucoux
- * 
- */
-
-/**
- * This object allows to add a ZAP command line option.
  *
- * @see <a href="https://code.google.com/p/zaproxy/wiki/HelpCmdline"> https://code.google.com/p/zaproxy/wiki/HelpCmdline</a>
  */
-public class ZAPCmdLine extends AbstractDescribableImpl<ZAPCmdLine> implements Serializable {
+public class ZAPReportCollection implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    @DataBoundConstructor
-    public ZAPCmdLine(String cmdLineOption, String cmdLineValue) {
-        this.cmdLineOption = cmdLineOption;
-        this.cmdLineValue = cmdLineValue;
+    private static ZAPReportCollection uniqueInstance = new ZAPReportCollection();
+
+    /** Map where key is the report format represented by a String and value is a ZAPreport object allowing to generate a report with the corresponding format. */
+    private Map<String, ZAPReport> mapFormatReport;
+
+    private ZAPReportCollection() {
+        mapFormatReport = new HashMap<String, ZAPReport>();
+
+        /* ZAPReport's creation. */
+        ZAPReportXML reportXML = new ZAPReportXML();
+        ZAPReportHTML reportHTML = new ZAPReportHTML();
+
+        /* Add ZAPReport to the map. */
+        mapFormatReport.put(reportXML.getFormat(), reportXML);
+        mapFormatReport.put(reportHTML.getFormat(), reportHTML);
     }
 
-    /** Configuration key for the command line */
-    private final String cmdLineOption;
+    public static ZAPReportCollection getInstance() {
+        if (uniqueInstance == null) uniqueInstance = new ZAPReportCollection();
 
-    public String getCmdLineOption() { return cmdLineOption; }
-
-    /** Configuration value for the command line */
-    private final String cmdLineValue;
-
-    public String getCmdLineValue() { return cmdLineValue; }
-
-    @Override
-    public String toString() {
-        String s = "";
-        s += "\n\t\tOPTION : [ " + cmdLineOption + " ]\n";
-        s += "\t\tVALUE  : [ " + cmdLineValue + " ]\n";
-        return s;
+        return uniqueInstance;
     }
 
-    @Extension
-    public static class ZAPCmdLineDescriptorImpl extends Descriptor<ZAPCmdLine> {
-
-        @Override
-        public String getDisplayName() {
-            return "ZAP Command Line";
-        }
-    }
+    public Map<String, ZAPReport> getMapFormatReport() { return mapFormatReport; }
 }
