+++ src/main/resources/com/github/jenkinsci/zaproxy/ZAPBuilder/global.jelly
