+++ src/main/resources/com/github/jenkinsci/zaproxy/ZAPDriver/help-loggedInIndicator.html
